Remove unnecessary type conversions

+++ src/irvine/oeis/a112/A112537.java
@@ -23,6 +23,6 @@ public class A112537 extends AbstractSequence {
 
   @Override
   public Z next() {
-    return mSeq1.next().pow(mSeq2.next().intValue());
+    return mSeq1.next().pow(mSeq2.next());
   }
 }
