+++ src/irvine/oeis/a112/A112538.java
@@ -24,6 +24,6 @@ public class A112538 extends AbstractSequence {
 
   @Override
   public Z next() {
-    return mSeq1.next().pow(mSeq2.next().intValue());
+    return mSeq1.next().pow(mSeq2.next());
   }
 }
