+++ src/irvine/oeis/a222/A222621.java
@@ -23,6 +23,6 @@ public class A222621 extends AbstractSequence {
 
   @Override
   public Z next() {
-    return mSeq1.next().pow(mSeq2.next().intValue());
+    return mSeq1.next().pow(mSeq2.next());
   }
 }
