+++ src/irvine/oeis/a275/A275465.java
@@ -24,6 +24,6 @@ public class A275465 extends AbstractSequence {
 
   @Override
   public Z next() {
-    return mSeq1.next().pow(mSeq2.next().intValue());
+    return mSeq1.next().pow(mSeq2.next());
   }
 }
