+++ src/irvine/oeis/a294/A294083.java
@@ -22,6 +22,6 @@ public class A294083 extends AbstractSequence {
 
   @Override
   public Z next() {
-    return mSeq1.next().pow(mSeq2.next().intValue());
+    return mSeq1.next().pow(mSeq2.next());
   }
 }
